+++ test/parallel/test-vm-proxies.js
@@ -16,16 +16,3 @@ sandbox = { Proxy: Proxy };
 vm.runInNewContext('this.Proxy = Proxy', sandbox);
 assert.strictEqual(typeof sandbox.Proxy, 'function');
 assert.strictEqual(sandbox.Proxy, Proxy);
-
-// Handle a sandbox that throws while copying properties
-assert.throws(() => {
-  const handler = {
-    getOwnPropertyDescriptor: (target, prop) => {
-      throw new Error('whoops');
-    }
-  };
-  const sandbox = new Proxy({foo: 'bar'}, handler);
-  const context = vm.createContext(sandbox);
-
-  vm.runInContext('', context);
-}, /whoops/);
